fix: extract the env variable to the settings.rs and some refactoring, and the trigger in windows-release.yml
