+++ .github/workflows/windows-release.yml
@@ -1,10 +1,11 @@
 name: Build Windows release
 
+
 on:
   workflow_dispatch:
   push:
     tags:
-      - 'v*'
+      - '[0-9]+.[0-9]+.[0-9]+'
     branches:
       - main
       - master
