+++ .gitignore
@@ -6,3 +6,5 @@
 .env
 
 .claude
+
+/docs
